Remove all uses of String#force_encoding
EventMachine::LineAndTextProtocol#receive_line sometimes hands us data
encoded as ASCII-8BIT (a raw byte string) instead of the character-
encoded string we expect. But forcing Ruby to transcode every string to
UTF-8 isn't the correct thing to do, especially when some of those
strings are frozen and can't be modified.

Instead, let's take a closer look at the ASCII-8BIT data that
EventMachine gives us--it may be the case that we're getting an
incomplete line, and need to wait for more data before acting on it.