Require clients to use a Unicode-compatible encoding
When Hector receives what appears to be a complete line from a client,
it will attempt to transcode the line to UTF-8, if it is not already. If
Ruby is unable to transcode the line, Hector will send
ERR_UNKNOWNERROR (400) to the client with the message "Please use
UTF-8", and then close the connection.