Don't attempt to transcode to UTF-8
EventMachine always gives us an ASCII-8BIT string. Even if the incoming
data is valid UTF-8, calling String#encode! on it raises an encoding
error. Instead of introducing more code paths, we'll always try forcing
the encoding to UTF-8. If it appears valid, great. If not, Hector will
send an error to the client and close the connection.